Modify playRound function, add game function
Moved the scoring function into the playRound function and changed the
returns in the playRound function into alerts.  I think the returns
were breaking the chain within the game function.  I used a for loop
within the game function and it now successfully plays through a game
of rock, paper, scissors until someone scores 5 points.

Changes to be committed:
	modified:   external.js